+++ api/admin.js
@@ -1,14 +1,3 @@
-function adminRequired(req, res, next) {
-  console.log(req.user.amininistrator);
-  if (!req.user.amininistrator === true) {
-    next({
-      name: "Unqualified",
-      message: "You cannot do this",
-    });
-  }
-  next();
-}
-
 const express = require("express");
 
 const adminRouter = express.Router();
@@ -17,5 +6,4 @@ adminRouter.use((req, res, next) => {
   console.log("Requesting Admin");
   next();
 });
-
-module.exports = { adminRouter, adminRequired };
+module.exports = adminRouter;
