+++ api/usersRouter.js
@@ -1,100 +1,102 @@
 const express = require("express");
 const jwt = require("jsonwebtoken");
 const {
-    getUser,
-    getAllUsers,
-    getUserByEmail,
-    getUserById,
-    createUser,
-    // deleteUser,
+  getUser,
+  getAllUsers,
+  getUserByEmail,
+  getUserById,
+  createUser,
+  // deleteUser,
 } = require("../db/index");
+
+const adminRequired = require("./utils");
 
 const usersRouter = express.Router();
 
-usersRouter.get("/all", async (req, res, next) => {
-    try {
-        const users = await getAllUsers();
-        res.send({ users });
-    } catch (error) {
-        next(error);
-    }
+usersRouter.get("/all", adminRequired, async (req, res, next) => {
+  try {
+    const users = await getAllUsers();
+    res.send({ users });
+  } catch (error) {
+    next(error);
+  }
 });
 
 usersRouter.get("/me", async (req, res, next) => {
-    try {
-        if (!req.user) {
-            res.status(400).send("No user has been found");
-        }
-        res.send(req.user);
-    } catch (error) {
-        next(error);
+  try {
+    if (!req.user) {
+      res.status(400).send("No user has been found");
     }
+    res.send(req.user);
+  } catch (error) {
+    next(error);
+  }
 });
 
 usersRouter.post("/login", async (req, res, next) => {
-    const { email, password } = req.body;
-    try {
-        const user = await getUser({ email, password });
+  const { email, password } = req.body;
+  try {
+    const user = await getUser({ email, password });
 
-        if (!user) {
-            res.send({
-                name: "invalidCredentials",
-                error: "Invalid email or password.",
-            });
-            return;
-        }
-
-        const token = jwt.sign(
-            {
-                email: user.email,
-                id: user.id,
-            },
-            process.env.JWT_SECRET
-        );
-        res.send({ user, token });
-    } catch (error) {
-        next(error);
+    if (!user) {
+      res.send({
+        name: "invalidCredentials",
+        error: "Invalid email or password.",
+      });
+      return;
     }
+
+    const token = jwt.sign(
+      {
+        email: user.email,
+        id: user.id,
+      },
+      process.env.JWT_SECRET
+    );
+    res.send({ user, token });
+  } catch (error) {
+    next(error);
+  }
 });
 
 usersRouter.post("/register", async (req, res, next) => {
-    const { email, password, first_name, last_name } = req.body;
-    const user = await getUserByEmail(email);
+  const { email, password, first_name, last_name } = req.body;
+  const user = await getUserByEmail(email);
 
-    if (user) {
-        res.send({ error: "User already exists." });
-        return;
-    }
+  if (user) {
+    res.send({ error: "User already exists." });
+    return;
+  }
 
-    try {
-        const newUser = await createUser({
-            email,
-            password,
-            first_name,
-            last_name,
-        });
+  try {
+    const newUser = await createUser({
+      email,
+      password,
+      first_name,
+      last_name,
+    });
 
-        const token = jwt.sign(
-            {
-                email: newUser.email,
-                id: newUser.id,
-            },
-            process.env.JWT_SECRET
-        );
-        res.send({ newUser, token });
-    } catch (error) {
-        next(error);
-    }
+    const token = jwt.sign(
+      {
+        email: newUser.email,
+        id: newUser.id,
+      },
+      process.env.JWT_SECRET
+    );
+    res.send({ newUser, token });
+  } catch (error) {
+    next(error);
+  }
 });
 
 usersRouter.get("/:id", async (req, res, next) => {
-    try {
-        const userId = req.params.id;
-        const user = await getUserById(userId);
-        res.send({ user });
-    } catch (error) {
-        next(error);
-    }
+  try {
+    const userId = req.params.id;
+    const user = await getUserById(userId);
+    res.send({ user });
+  } catch (error) {
+    next(error);
+  }
 });
 
 // usersRouter.delete("/delete/:id", async (req, res, next) => {
