+++ src/components/index.js
@@ -4,4 +4,3 @@ export { default as Navbar } from "./navbar";
 export { default as Login } from "./login";
 export { default as Register } from "./register";
 export { default as Info } from "./info";
-export { default as Admin } from "./admin";
